backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
python-magic-bin==0.4.14
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Settings and Environment
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Audio Processing
SpeechRecognition==3.10.0
pydub==0.25.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
openai-whisper==20231117
ffmpeg-python==0.2.0
soundfile==0.12.1
librosa==0.10.1

# AI/ML
google-generativeai==0.3.1
numpy==1.26.2
scikit-learn==1.3.2
transformers==4.36.1
--find-links https://download.pytorch.org/whl/torch_stable.html
sentencepiece==0.1.99

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0
faker==21.0.0

# Process Management
psutil==5.9.7
